# docs/openapi.yaml
openapi: 3.1.0
info:
  title: "ZarishSphere FHIR Engine API"
  description: |
    FHIR R5 REST API server for ZarishSphere healthcare platform.
    
    ## Authentication
    All endpoints require a valid SMART on FHIR 2.1 JWT Bearer token.
    See https://docs.zarishsphere.com/auth for details.
    
    ## Versioning
    API version is embedded in the URL: `/fhir/R5/` (FHIR version, not service version).
    
    ## Multi-Tenancy
    All requests are tenant-isolated. The tenant ID is extracted from JWT claims.
  version: "1.0.0"
  contact:
    name: "ZarishSphere Platform"
    email: "platform@zarishsphere.com"
    url: "https://zarishsphere.com"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0"
servers:
  - url: "https://api.zarishsphere.com"
    description: "Production"
  - url: "http://localhost:8001"
    description: "Local development"
security:
  - bearerAuth: []
paths:
...
  /fhir/R5/{resourceType}/{id}/_history:
    get:
      summary: "Get resource history"
      operationId: "getHistory"
      tags: ["History"]
      parameters:
        - name: "resourceType"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "_since"
          in: "query"
          schema:
            type: "string"
        - name: "_until"
          in: "query"
          schema:
            type: "string"
      responses:
        "200":
          description: "Bundle of history entries"
  /fhir/R5/{resourceType}/{id}/_history/{versionId}:
    get:
      summary: "Get specific version"
      operationId: "getVersion"
      tags: ["History"]
      parameters:
        - name: "resourceType"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "versionId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "Specific version of resource"

  # Analytics Dashboard
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
...